+++ cmd/frontend/internal/app/router/router.go
@@ -31,6 +31,8 @@ const (
 
 	RegistryExtensionBundle = "registry.extension.bundle"
 
+	UsageStatsDownload = "usage-stats.download"
+
 	OldToolsRedirect = "old-tools-redirect"
 	OldTreeRedirect  = "old-tree-redirect"
 
@@ -83,6 +85,8 @@ func newRouter() *mux.Router {
 
 	addOldTreeRedirectRoute(base)
 	base.Path("/tools").Methods("GET").Name(OldToolsRedirect)
+
+	base.Path("/site-admin/usage-statistics/archive").Methods("GET").Name(UsageStatsDownload)
 
 	if envvar.SourcegraphDotComMode() {
 		base.PathPrefix("/go/").Methods("GET").Name(GoSymbolURL)
