+++ cmd/frontend/internal/httpapi/httpapi.go
@@ -76,9 +76,6 @@ func NewHandler(m *mux.Router, schema *graphql.Schema, githubWebhook, bitbucketS
 	m.Get(apirouter.SrcCliVersion).Handler(trace.TraceRoute(handler(srcCliVersionServe)))
 	m.Get(apirouter.SrcCliDownload).Handler(trace.TraceRoute(handler(srcCliDownloadServe)))
 
-	// Usage statistics ZIP download
-	m.Get(apirouter.UsageStatsDownload).Handler(trace.TraceRoute(handler(usageStatsDownloadServe)))
-
 	m.Get(apirouter.Registry).Handler(trace.TraceRoute(handler(registry.HandleRegistry)))
 
 	m.NotFoundHandler = http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
