+++ cmd/frontend/internal/httpapi/router/router.go
@@ -13,8 +13,6 @@ const (
 	SrcCliVersion  = "src-cli.version"
 	SrcCliDownload = "src-cli.download"
 
-	UsageStatsDownload = "usage-stats.download"
-
 	Registry = "registry"
 
 	RepoShield  = "repo.shield"
@@ -71,7 +69,6 @@ func New(base *mux.Router) *mux.Router {
 	base.Path("/lsif/upload").Methods("POST").Name(LSIFUpload)
 	base.Path("/src-cli/version").Methods("GET").Name(SrcCliVersion)
 	base.Path("/src-cli/{rest:.*}").Methods("GET").Name(SrcCliDownload)
-	base.Path("/usage-stats/archive").Methods("GET").Name(UsageStatsDownload)
 
 	// repo contains routes that are NOT specific to a revision. In these routes, the URL may not contain a revspec after the repo (that is, no "github.com/foo/bar@myrevspec").
 	repoPath := `/repos/` + routevar.Repo
